OOP code, better error handling, allow local testing
Improve error handling by using object-oriented code with destructors
and by catching probable exceptions when processing files so that other
files are still rotated if one file fails for whatever reason.

Make the domain configurable and avoid direct database access so that
the code can be run locally, potentially in a test wiki instead of live
Commons. The pages used (category, template, user subpages) are still
hardcoded, but they can be copied to the test wiki to make testing
possible.