Remove concurrency and process handling
The lock file is unnecessary: the Toolforge jobs framework ensures
(since [1]) that a cron job fails to start if another instance of the
same job is running.

This also means that `killAllRotatebots` makes no sense anymore: if an
instance is stuck, the killer instance won’t start; if the killer
instance can start, there is nothing to clean up. (This means that
interface admins can no longer do anything if the bot gets stuck, but
hopefully this won’t happen. If it happens, a maintainer of the tool
or – I believe – a Toolforge admin can kill it.)

[1] https://gitlab.wikimedia.org/repos/cloud/toolforge/jobs-api/-/commit/6fd890a4450eb7637d1cd4ff76c97c1854d08678